Dockerfile: Use docker to build this project and its subsidiary WAR
It's a bit of a pain to build this project, so let's have docker do it
for us. First install some basic dependencies for opentsdb, then clone
the project, check out the proper version, and build it. Then do it
again for tsquare itself. Note the CMD at the end of this file: Invoking
`docker build` on an image produced by this Dockerfile will output the
tsquare.war to stdout. In other words, on a docker-configured system,
you can generate the tsquare WAR with two commands:

    $ docker build -t="tsquare:v1" .
    $ docker run -i -t tsquare:v1 > tsquare.war